requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
celery==5.4.0
celery[redis]==5.4.0
celery-beat==2.5.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
